// src/server/services/twilio/twilio.schema.ts
export interface TwilioMessage {
  MediaContentType0?: 'audio/ogg';
  MediaUrl0?: string;
  SmsMessageSid: string;
  NumMedia: string;
  ProfileName: string;
  MessageType: string;
  SmsSid: string;
  WaId: string;
  SmsStatus: string;
  Body: string;
  To: string;
  NumSegments: string;
  ReferralNumMedia: string;
  MessageSid: string;
  AccountSid: string;
  From: string;
  ApiVersion: string;
}